Revert "Requires extended regular expressions to find jar (#13)"
This reverts commit 597ddff8bd84209d370d73bb1733f921ec0d28ad.

This PR was created to fix a bug that no longer exists.  The solution of
using extended regular expressions was in fact incorrect, but just
happened to cancel out the bug in certain cases.  Now that the bug has
been fixed (dabf132), this breaks the regex in all cases.

Fixes #16