Fix broken `ExecStart` command (#18)
* Revert "Requires extended regular expressions to find jar (#13)"

This reverts commit 597ddff8bd84209d370d73bb1733f921ec0d28ad.

This PR was created to fix a bug that no longer exists.  The solution of
using extended regular expressions was in fact incorrect, but just
happened to cancel out the bug in certain cases.  Now that the bug has
been fixed (dabf132), this breaks the regex in all cases.

Fixes #16

* Break long command

Improves readability and version control

* Find .jar file using `find`

The output of `ls` cannot be relied upon.  `find` is a much more robust
scripting tool.

* Simplify regex

* Use `xargs -0` instead of parameter substitution

More robust against filenames containing whitespace characters